[Fixes bug #747607] Make the Color Range slider work for Pencil & Ink Sketch
Since the BrightnessContrastEffect.Calculate effect is somewhat
expensive, it is not computed on every call to Render (only when
the table_calculated member is false). This meant that any adjustments
to the brightness data after the call to Calculate() had no effect on
the rendered image, since table_calculated remained the same.

Solution: set table_calculated to false whenever any properties
of the BrightnessContrastData class are modified