[Fixes bug #730054] Fix incorrect zoom text validition when focus is lost
When the zoom combo box loses focus, the zoom value is validated and
updated. This validition was incorrect (didn't remove the trailing percent
sign) so the zoom value was reset.

I created a new method (TryParsePercent) to perform the zoom text validation
(which is done in a couple other places) to remove duplicated code and
prevent similar bugs in the future.