[Fixes bug #633104] Fix broken shortcuts when Num Lock or Caps Lock is enabled
Num Lock and Caps Lock appear as active modifier keys if they are enabled,
which makes it difficult to correctly check if more important keys like Ctrl and
Shift are active or not.

Added a new function FilterModifierKeys that filters out all modifier
keys except for Ctrl/Shift/Alt. This function also simplifies the bug
fix for issue #852770.