Added ability to remove tools
Now add-ins can clean up properly by removing any tools they've added.
So now the core tools package can be disabled and remove all of your
tools! (Probably not very useful, but important as a principle.) Likely
there might be other functionality broken because it assumes the existence of
some tool, but this we don't care about because who would want to do it?

Next up is removing effects.