Fixed LassoSelectTool.
When the lassoPolygon was being cleared after the mouse was released
(and after it was added to the doc.Selection.SelectionPolygons
collection), the Polygon inside SelectionPolygons was also being cleared
since it was only referencing to lassoPolygon rather than being a copy.
I fixed this by including System.Linq and then calling the ToList method
on lassoPolygon to make a complete copy of the Polygon.