Optimized calculation of elliptical curves.
The CalculateCurvePoints method was reduced to about 46.46% of the time
(ticks) it took for the old calculation code to execute, on average.
This can be multiplied by 4 for each curve of an ellipse and then by the
number of ellipses that are calculated when the mouse is being dragged.
Overall, it makes the elliptical calculations much faster.