Added an encapsulating UserLayer class.
UserLayer is a child of the Layer class, and in the future it will
encapsulate some other properties as well. UserLayer represents the
Layers that a user of Pinta can interact with directly, so wherever this
was the case, a Layer was replaced by a UserLayer. This was done for
adding re-editable text in the near future, as every UserLayer will need
to contain a class that has some properties for editable text *as well
as a Layer*.