Added a TextLayer to every UserLayer instance.
For every UserLayer instance there is a TextLayer that belongs to it.
All text (that's not being edited) will now be drawn on the TextLayer
rather than the UserLayer - the only exception to this is when text is
being edited, upon which it will be drawn to the ToolLayer like before.
This is subject to change in the future.