[Fixes bug #956195] Fix broken effects due to a change in Gtk# 2.12.11.
The implementations of Rectangle.Right and Rectangle.Bottom were
changed, which caused most effects to do nothing if a rectangle had
height 1 (in particular, the AsyncEffectRender class generates tiles of
that height).

I added a GetBottom function and a GetRight function that use the new
implementation, since that will make the eventual transition easier once
all distros have the latest version.