[Fixes bug #996490] Only create a single history item when pasting.
Before, if the canvas was expanded when pasting, two history items
would be created (one for resizing the canvas, and one for pasting).
As a result, when trying to undo the paste, the user had to press
Undo twice to fully undo the action.

This change uses a CompoundHistoryItem to merge the two history items
into a single history item.