[Fixes bug #987641] Fix reading of TGA files created by Pinta on Linux.
GDK seems to identify TGA files as ICO files if the image ID field is
empty (i.e. the first byte is zero). This change adds some content to
the image ID field so that Pinta can successfully open TGA files that
were created by it.

This is a bit of a hack, but it is allowed by the TGA spec and the
resulting files are still valid TGA files that can be opened by other
applications.