[Fixes bug #988278] Improve the error dialog when opening a file fails.
We now show an error message of "Unrecognized image file format" if the
image format is unknown, and an error message of "Permission denied" if
the user tries to open a file that they do not have permissions for.

The styling of the dialog was also changed to make it more consistent
with other dialogs in Pinta.