[Fixes bugs #724982, #606240] Fix issues with partially missing selections.
The canvas offset was being used in addition to the layer offset when
rendering a selection that was translated by the Move Pixels tool.

This causes issues once the canvas is zoomed in and the canvas offset
is non-zero, because the selection appears to be shifted and partially
(or entirely) missing.