Improve the performance of the SurfaceDiff algorithm.
Part of the algorithm is now run in parallel, and it allocates much
less memory in most situations. Additionally, it does less work when
computing the bounding box of the changed pixels. In my testing,
these changes made the algorithm anywhere from 2 to 15 times faster,
depending on the size of the changed pixels relative to the size of
the image.