Fix for 73b856af8560e2dbb340bac8e1547317a8b0f625.
The previous fix caused severe performance regression when drawing in the right or bottom half of the canvas because it was always triggering a full invalidate. (http://i.imgur.com/GWLNK.png)

I think this fixes the disappearing canvas issue (bug #1015345) without the performance loss by clamping the invalidate rectangle to the canvas.