Use ZipLib v. 2.84 to work with Mono.Addins on Win
Pinta does not require a specific version of ZipLib, but the Mono.Addins
DLL we bundle does. On Linux systems this does not become a problem since
they're all fetched from the GAC and work nicely together, but on Windows
we used the bundled DLLs and it turns out that Mono.Addins couldn't handle
the version we bundled. This was only discovered when we actaully tried to
use Mono.Addins on a mpack that had to be unzipped.

Taking ZipLib v 2.84 from Ubuntu satisfies Mono.Addins and works for Pinta
as well, meaning add-in installs on Windows work properly. (In theory,
there are other issues as well.) Since Linux uses the GAC, there should be
no difference on that platform. Not sure about Mac, but I'll wager they
either had the same problem as Windows or aren't affected like Linux.