[Fixes bug #1186871] Fix several layout issues with the text tool.
The bounding rectangle returned by Pango is much larger than what we
need, and the code previously tried to make some adjustments. However,
that didn't work very well, especially with large font sizes.

Now, we just set the height to be the # of lines times the cursor
height, which works very well.