(Terrible state) Implemented basics of editable ellipses and rounded shapes.
Warning: this commit leaves some tools in an awful,
alpha-development-stage state, including the Rectangle, Ellipse, Rounded
Rectangle, and some of the Select tools. On the plus side, the basic
functionality of editable ellipses and rounded shapes (beyond just 4
corners) has been added, and the rest of Pinta seems to be working okay!