Add new add-in repository, replacing wonky old one
The new repository is much simpler and more robust, if maybe slightly
less elegant. Also, it will no longer cost me money, which I see as a
significant performance optimisation.

The new repository requires add-in authors to clone the community addin
repository and add their addin there, which is about as much work for them
as registering on the Cydin site we tried to use before.

Also, there is currently no division between platforms in the new repository.
This might be a problem, but I didn't like the extra complexity, and to be
fair I think platform-dependence problems are something add-in authors should
tend to, not us.