Fix #1361560 - Fix many issues with the text engine.
The main change is to store the start position of the selection, instead
of a relative offset from the current cursor - this simplifies the code
quite a bit when the cursor is moving around.

The PositionToIndex/IndexToPosition calls in several areas (handling the
home/end keys, displaying selections, and deleting selections) were
completely wrong and did not work for Unicode text.