Fix #1422710, 1422715 - Fix use of Recolor tool after use of Editable shape.
The issue is that Recolor temporarily draws on the ToolLayer, which it expects to be invisible.  Editable shapes make the ToolLayer visible when activated but never hides it, so the Recolor tool's temporary works is visible.  Fix by hiding the tool layer when an editable tool finalized all its shapes.  Doing it on tool deactivation misses a case where you switch to a new image and then change the tool.