Initial work on a documentation system for Pinta.
This uses Mallard (http://projectmallard.org/), which is fairly standard
for Gnome applications and integrates with the help viewer, but will
also allow us to generate HTML help for other platforms and the website.

Mallard is aimed at simple, topic-oriented help pages, which I think is
a good match for Pinta.

The source documents are located under help/C/*.page, and running 'make
updatepot' generates a translation template that is used to produce
translated versions of the help pages at build time.

Bug: #778981