Source Operator not correctly set in LivePreviewManager.cs
changing ctx.Operator before a call to layer.Draw(...) has no effect because the BlendMode is overwritten by layer.Draw(...).
The error can be seen e.g. when having a colored semi-transparent image and apply a black the white filter. In the regions of the image where alpha is != 255 there remain colored parts.
Note that this is a REALLY REALLY DIRTY FIX for the problem. It should be discussed how to do it properly, because I am not completely sure how to do it best...