Update es.po - text breaks functionality
The current had an illegal character (0xC2) that breaks zooming, so that it just doesnt' work (neither Ctrl++ nor the buttons) when "es" is the current locale. See for example: https://bugs.launchpad.net/pinta/+bug/1464855

This fixes that by turning it into a space, that probably was what the translator meant.