Fix issues where the document's dirty state was not updated properly.
The history stack should track where the last clean item was, so that it is marked as modified when switching away from that item, and unmodified when returning to that item.

This also avoids the incorrect assumption that returning to the first history item clears the modified state, which is wrong if the document was saved at a later history item.

Fixes: #1905165