Fix a bug where deselecting via the select tool could cause undo-related issues
Since the OnMouseDown handler resets some of the selection data members, this should be reverted before activating the deselect tool, which grabs a snapshot of the current selection.

In the referenced bug report, this eventually caused the image to be "lost" because the Finish Pixels action was invoked with a selection that had nothing selected.

Fixes: #1905719