Cherry-pick 7bd148fe from the GTK3 branch.
 Fix a couple issues with 3915f8e:

When typing in the text tool and pressing 'X', this would also be handled by the palette instead of letting the key press be consumed by the text tool.
Also, pressing X didn't have any effect when there weren't any open documents, since it was being handled by the canvas widget.

The key event is now handled on key down, for consistency with the shortcuts for switching between tools.