Rework how the selection move handles are rendered.
Previously these were rendered on the ToolLayer, which had inherent
limitations from being tied to the image size. For example, this made it
difficult to perform selections when zoomed in on small images since
each handle took up multiple pixels, and zooming in after making a
selection caused the handles to be zoomed as well.

The handles are now drawn over the canvas at a consistent size on the
screen. There is a new hook (IToolHandle) to allow tools to provide
handles that can render themselves on the canvas. This has a fairly
minimal API, so it's still up to tools to handle mouse events, decide
when to redraw, etc similar to how tools work already

This also turned into a larger refactoring of the SelectTool. In
particular, the Shift key constraint to draw square selections was
reimplemented in three(!) different places.
This was removed from Utility.PointsToRectangle, which also was
introducing an extra padding of 1 pixel that was noticeable when drawing
selections on small images

The select tool also now rounds the handle points to the nearest image pixel.
This behaves much better when zoomed in on a small image, where it looks
confusing if the move handles are at a subpixel location.

Fixes: 1173756