Initial work on using Gio.File for dealing with files
This is a step towards Gtk 4, where the local-only setting for the file chooser no longer exists. Using gio.File will allow us to better handle virtual filesystems: for example, in the scenario for bug 1958763 we can now access the user-friendly display name for the file

This commit uses Gio.File for opening images and tracking the last dialog directory.

Bug: 1958763