Avoid generating zero-width dashes for Cairo
When the line cap is square this still draws a single square with the brush width, so this isn't a no-op like the butt line cap type.
Instead, we avoid this by cycling the space to the end of the pattern and then offsetting where the pattern starts drawing from.

Bug: 1959032