Rework drag&drop handling to use the simpler DropTarget method
This returns a Gdk.FileList with the Gio.File for each item that was dropped. The previous approach worked, but was hitting some mysterious internal crashes in GTK, while this new method works much more reliably in my testing

Dropping images from a web browser doesn't work at all on the macOS backend for GTK, so this needs testing on other platforms.