Refactor some of the selection tool's code into a reusable rectangle handle
The goal is to allow this to be used in the Move Selected Pixels tool for scaling the selection in a more intuitive manner

The Select tool's behaviour should be unchanged, except for one minor difference. If the rectangle is inverted (e.g. dragging the bottom right corner before the top left corner), after a mouse up the rectangle resets itself to no longer be inverted so that the mouse cursors for each handle point in the expected direction

Bug: 1995636