Add methods for decomposing a Cairo.Matrix
This will be useful for the Move tool to be able to align the handles with the current transform

Also fixed an issue where modules like Cairo were uninitialized during unit tests unless a method from the Utilities class was called. This is now marked with the correct attributes to run before any unit test.