Generate output files under a `build` folder, rather than just `bin`
For Windows the plan is to copy the MSYS GTK libs to the build folder so Pinta can be easily launched without manipulating the path, so this is most easily done by adding sibling `lib` and `share` folders.

This doesn't affect the output of 'dotnet publish', so the existing Linux and macOS install scripts should still work.