Remove useless code (#1366)
The style class does nothing in this case.
The height request is also useless for the toolbox. There should be some height request some ware to allow the history undo and redo buttons to be always visible, but this is probably not the best place and is also insufficient.

(cherry picked from commit 811ae97310902c391d26a31ccabd96d9203d7dd7)