Initialize the application menubar during the 'startup' event
When testing the non-headerbar layout on Linux (KDE), doing this in the 'activate' event caused the menubar to not appear - it seems necessary to do this during the startup event. This also seems to have fixed intermittent issues on macOS where the global menu was left empty.

The header bar is still populated during the 'activate' event along with creating the rest of the window, so some refactoring was needed to enable this.