Add a flag to prevent infinite loop if changing layers adds a history item
When a history item is added, the list view is rebuilt and the selection is reset back to its original value. When this happens, we don't need to trigger another change to the document's active layer since the selection won't be any different. This previously led to an infinite loop because the current tool is committed before updating the document's active layer.

Fixes: #1463